refactor: move GeminiConfig definition from internal/config to internal/gemini package
